requirements.txt
flask
flask_httpauth
flask-limiter
pymemcache
pillow
# Instalar sudo apt install memcached

